书籍部分完成

+++ backend/app/Http/Controllers/BookManageController.php
@@ -143,6 +143,11 @@ class BookManageController extends Controller
 			foreach($orders as $order){
 				$order->delete();
 			}
+			$bookimages = \App\Model\BookImage::where('book_id', $info->id)
+											 ->get();
+			foreach($bookimages as $bookimage) {
+				$bookimage->delete();
+			}
 			$res = $info->delete();
 			if($res){
 				return response()->json(['delete' => true], 200);
@@ -154,27 +159,73 @@ class BookManageController extends Controller
 		}
     }
     // 更新
-    public function updateBook(Request $request){
-        $book = \App\Model\Book::find($request->id);
-        if($book){ 
-            $book2 = \App\Model\Book::where('bookcode', $book->bookcode)
-                        ->first();
-            if($book2->id != $request->id){
-                return response()->json(['update' => false, 'err' => "this bookcode already exist"], 200);
+    public function updateBookISBN(Request $request){
+        $book = \App\Model\Book::where('id', $request->id)
+								->first();
+		if($book){			
+			if($request->category){
+                $book->category = $request->category;
+            } else {
+				return response()->json(['create' => false], 200);
+                $book->category = '';
             }
-            $book->bookname = $request->bookname;
-            $book->bookcode = $request->bookcode;
-            $book->description = $request->description;
-            $book->category_id = $request->category_id;
-            $book->ISBN = $request->ISBN;
-            $book->bookpublishtime = $request->bookpublishtime;
-            $book->booklocation = $request->booklocation;
-            $book->bookauthor = $request->bookauthor;
-            $book->save();
-            return response()->json(['update' => true], 200);
-        }else{
-            return response()->json(['update' => false, 'err' => 'Not Exist This Id'], 200);
-        }
+            if($request->title) {
+                $book->title = $request->title;
+            } else {
+				return response()->json(['create' => false], 200);
+                $book->title = '';
+            }
+            if($request->author) {
+                $book->author = $request->author;          
+            } else {
+                $book->author = '';                        
+            }
+            if($request->translator) {
+                $book->translator = $request->translator;               
+            } else {
+                $book->translator = '';                                               
+            }
+            if($request->publisher) {
+                $book->publisher = $request->publisher;                                        
+            } else {
+                $book->publisher = '';                                                               
+            }
+            if($request->pudate){
+                $book->pudate = $request->pudate;
+            } else {
+                $book->pudate = '';
+            }
+            if($request->authorintro){
+                $book->authorintro = $request->authorintro;
+            } else {
+                $book->authorintro = '';
+            }
+            if($request->summary){
+                $book->summary = $request->summary;
+            } else {
+                $book->summary = '';                
+            }
+            if($request->catalog){
+                $book->catalog = $request->catalog;
+            } else {
+                $book->catalog = '';       
+            }
+            if($request->price){
+                $book->price = $request->price;
+            } else {
+                $book->price = '';
+            }
+            if($request->pages){
+                $book->pages = $request->pages;
+            } else {
+                $book->pages = '';
+            }
+			$book->save();
+			return response()->json(['update' => true], 200);
+		} else{
+			
+			return response()->json(['update' => false], 200);
+		}
     }
     public function getAllBook(){
         $books = \App\Model\Book::all()
@@ -186,13 +237,75 @@ class BookManageController extends Controller
 								->get();
 		foreach($books as $book) {
 			$orders = \App\Model\Order::where('book_id', $book->id)
-									  ->first();
+									  ->get();
+			foreach($orders as $order) {
+				$order->delete();
+			}
+			$bookimages = \App\Model\BookImage::where('book_id', $book->id)
+											 ->get();
+			foreach($bookimages as $bookimage) {
+				$bookimage->delete();
+			}
+			$book->delete();
+		}
+		return response()->json(['delete' => true], 200);
+	}
+	public function	updateBooklocation(Request $request) {
+		$book = \App\Model\Book::where('id', $request->id)
+								->first();
+		if($book){
+			$book->location = $request->location;
+			$book->save();
+			return response()->json(['update' => true], 200);
+		} else {
+			return response()->json(['update' => false], 200);
+		}
+	}
+	
+	public function bindBookImage(Request $request) {
+		$books = \App\Model\Book::where('isbn', $request->isbn)
+								->get();
+		foreach($books as $book){
+			$image = new \App\Model\BookImage();
+			$image->image_id = $request->imageid;
+			$image->book_id = $book->id;
+			$image->save();
+		}
+		return response()->json(['bind' => true], 200);
+	}
+	
+	public function getBookAllImage(Request $request) {
+		$book = \App\Model\Book::where('isbn', $request->isbn)
+								->first();
+		if($book){
+			$images = \App\Model\BookImage::where('book_id', $book->id)
+											->get();
+			return response()->json(['get' => true, 'images' => $images], 200);
+		}else {
+			return response()->json(['get' => false], 200);
+		}
+	}
+	public function CancelBindBookImage(Request $request) {
+		$bookimages = \App\Model\BookImage::where('image_id', $request->imageid)
+										  ->get();
+		foreach($bookimages as $bookimage) {
+			$bookimage->delete();
+		}
+		return response()->json(['cancel' => true], 200);
+	}
+	/* public function deleteBookById(Request $request){
+		$book = \App\Model\Book::where('id', $request->id)
+								->first();
+		if($book) {
+			$orders = \App\Model\Order::where('id', $book->id)
+									->get();
 			foreach($orders as $order) {
 				$order->delete();
 			}
 			$book->delete();
+			return response()->json(['delete' => true], 200);
+		} else {
+			return response()->json(['delete' => true], 200);
 		}
-		response()->json(['delete' => true], 200);
-	}
-	
+	} */
 }
